[feat]peacock page
