+++ client/src/Page/peacockDetail.tsx
@@ -1,0 +1,3 @@
+const PeacockDetail = () => {
+  return <div></div>;
+};
